refactor(auth): move auth logic to components folder
