+++ FinEdgeFrontend/src/Pages/Register.tsx
@@ -17,7 +17,6 @@ const Register = () => {
   const queryClient = useQueryClient();
   const emailRegEx = new RegExp("^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$");
   const passwordRegEx = new RegExp("^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{10,}$");
-
   const [emailError, setEmailError] = useState<boolean>(true);
   const [passwordError, setPasswordError] = useState<boolean>(true);
   const [isNextButtonDisabled, setIsNextButtonDisabled] = useState<boolean>(true);
